improve magic load
