+++ apps/next/src/pages/_app.tsx
@@ -5,7 +5,6 @@ import "setimmediate";
 import "@rainbow-me/rainbowkit/styles.css";
 import { AppProps } from "next/app";
 import Head from "next/head";
-import Script from "next/script";
 
 import { View } from "@showtime-xyz/universal.view";
 
@@ -13,7 +12,6 @@ import { Footer } from "app/components/footer";
 import { Header } from "app/components/header";
 import { withColorScheme } from "app/components/memo-with-theme";
 import { useLogRocket } from "app/hooks/use-logrocket";
-import { onMagicLoad } from "app/lib/magic/magic-load-listener";
 import { renderEmptyAnalyticsSnippet } from "app/lib/rudderstack/script";
 import { Sentry } from "app/lib/sentry";
 import { AppProviders } from "app/providers/app-providers";
@@ -113,11 +111,6 @@ export default function App({ Component, pageProps, router }: AppProps) {
           dangerouslySetInnerHTML={{ __html: renderEmptyAnalyticsSnippet() }}
         />
       </Head>
-      <Script
-        src="https://cdn.jsdelivr.net/npm/magic-sdk/dist/magic.js"
-        strategy="afterInteractive"
-        onLoad={onMagicLoad}
-      />
       <AppProviders>
         <Container>
           <Header
