+++ packages/app/lib/magic/magic.web.ts
@@ -1,0 +1,7 @@
+import { Fragment } from "react";
+
+import { useMagic } from "app/providers/magic-provider.web";
+
+const Relayer = Fragment;
+
+export { useMagic, Relayer };
